refactor(bus): remove unused /bus page with BusSearchInput
